[chore] App: set up basic DI and move files to correct packages
