+++ composeApp/src/androidMain/kotlin/com/rwmobi/roctopus/MainActivity.kt
@@ -1,6 +1,5 @@
 package com.rwmobi.roctopus
 
-import App
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
