+++ composeApp/src/androidMain/kotlin/com/rwmobi/roctopus/Platform.android.kt
@@ -1,3 +1,12 @@
+/*
+ * Copyright (c) 2024. Ryan Wong
+ * https://github.com/ryanw-mobile
+ * Sponsored by RW MobiMedia UK Limited
+ *
+ */
+
+package com.rwmobi.roctopus
+
 import android.os.Build
 
 class AndroidPlatform : Platform {
