+++ composeApp/src/commonMain/kotlin/com/rwmobi/roctopus/di/AppModule.kt
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) 2024. Ryan Wong
+ * https://github.com/ryanw-mobile
+ * Sponsored by RW MobiMedia UK Limited
+ *
+ */
+
+package com.rwmobi.roctopus.di
+
+import org.koin.dsl.module
+
+val appModule = module {
+    // placeholder
+}
