+++ composeApp/src/desktopMain/kotlin/com/rwmobi/roctopus/Platform.jvm.kt
@@ -1,3 +1,12 @@
+/*
+ * Copyright (c) 2024. Ryan Wong
+ * https://github.com/ryanw-mobile
+ * Sponsored by RW MobiMedia UK Limited
+ *
+ */
+
+package com.rwmobi.roctopus
+
 class JVMPlatform : Platform {
     override val name: String = "Java ${System.getProperty("java.version")}"
 }
