--- composeApp/src/desktopMain/kotlin/main.kt
@@ -1,11 +1,0 @@
-import androidx.compose.ui.window.Window
-import androidx.compose.ui.window.application
-
-fun main() = application {
-    Window(
-        onCloseRequest = ::exitApplication,
-        title = "Roctopus",
-    ) {
-        App()
-    }
-}
