--- composeApp/src/iosMain/kotlin/MainViewController.kt
@@ -1,3 +1,0 @@
-import androidx.compose.ui.window.ComposeUIViewController
-
-fun MainViewController() = ComposeUIViewController { App() }
