+++ composeApp/src/iosMain/kotlin/com/rwmobi/roctopus/Platform.ios.kt
@@ -1,3 +1,12 @@
+/*
+ * Copyright (c) 2024. Ryan Wong
+ * https://github.com/ryanw-mobile
+ * Sponsored by RW MobiMedia UK Limited
+ *
+ */
+
+package com.rwmobi.roctopus
+
 import platform.UIKit.UIDevice
 
 class IOSPlatform : Platform {
